AHN.Buildings.Parallel: updated to handle GeoTiff files as input, both for AHN-2 and AHN-3. Added file existence checking.
AHN.Buildings: fixed intermediate result file deletion in FILES mode.

+++ AHN.Buildings/Result.cpp
@@ -42,7 +42,16 @@ Result& Result::operator=(Result&& other)
 
 TemporaryFileResult::~TemporaryFileResult()
 {
-	fs::remove(_path);
+	if (!_path.empty())
+	{
+		if (dataset)
+		{
+			// necessary, because parent dtor will be called after this
+			GDALClose(dataset);
+			dataset = nullptr;
+		}
+		fs::remove(_path);
+	}
 }
 
 #pragma endregion
